test(audit): move reset_dataset tests to backend-agnostic file
Consolidates xtdb + mariadb reset_dataset cases into
tests/backends/test_audit_reset_dataset.py. Shared connection/orm stubs
factored to module-scope. test_xtdb_audit_ops.py keeps only xtdb-specific
tests.

WG-272
